requirements.txt
# OpenAI API 依赖
openai>=1.0.0

# Web服务器依赖
flask>=2.0.0
flask-cors>=3.0.0

# 其他可能需要的依赖
requests>=2.28.0
python-dotenv>=1.0.0